# OTT-Platform-Web-Application/Backend/docs/api_docs.yaml
...
openapi: 3.0.3
info:
  title: Dhaakad API
  description: Dhaakad API documentation (Might be buggy)
  version: 0.0.5
  contact:
    email: vishal.rustagi@metaorangedigital.com
servers:
  - url: 'https://api.dhaakadcinema.com/api/v1'
...
paths:
...
  /users/resetPassword/{id}:
    patch:
      tags:
        - users
      summary: reset password
      operationId: resetPassword
      parameters:
        - name: id
          in: path
          description: user id
          required: true
          schema:
            type: string
            example: 'abc123'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: 'abcd'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Password changed successfully!"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /users/verifyOTP:
    patch:
      tags:
        - users
      summary: verify OTP
      operationId: verifyOTP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: 'sandy@gmail.com'
                otp:
                  type: integer
                  example: 123456
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: 'abc123'
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "OTP verified successfully"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/resendOTP:
    post:
      tags:
        - users
      summary: resend OTP
      operationId: resendOTP
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: 'sandy@gmail.com'
      responses:
        200:
          description: OTP sent to email
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "OTP sent to email!"
        400:
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /casts:
    get:
      tags:
        - casts
      summary: get list of casts details
      operationId: getListOfCasts
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "casts get successfully"
                  results:
                    type: number
                    example: 10
                  data:
                    type: object
                    properties:
                      cast:
                        type: array
                        items:
                          $ref: '#/components/schemas/Casts'
    post:
      tags:
        - casts
      summary: create Cast
      operationId: createCast
      requestBody:
        description: upload casts object
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/Casts'
                - type: object
                  properties:
                    photo:
                      type: string
                      format: 'file upload'
      responses:
        200:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "casts get successfully"
                  results:
                    type: number
                    example: 10
                  data:
                    type: object
                    properties:
                      cast:
                        $ref: '#/components/schemas/Casts'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    User:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'asdasd'
        firstName:
          type: string
          example: 'Sandy'
        lastName:
          type: string
          example: 'Joseph'
        userName:
          type: string
          example: 'Sandy'
        email:
          type: string
          example: 'sandy@gmail.com'
        referralCode:
          type: string
          example: 'test'
        idReferralCode:
          type: boolean
          example: true
        rewards:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
                example: 'Referral Reward'
              amount:
                type: number
                example: 10
        coupons:
          type: array
          items:
            type: string
          example: ['abc','xyz']
        otp:
          type: number
          example: '123456'
        expiryotp:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        password:
          type: string
          example: 'abcd'
        passwordConfirm:
          type: string
          example: 'abcd'
        mobile:
          type: 'string'
          example: '+91-1234567891'
        gender:
          type: string
          example: 'male||female'
        dateOfBirth:
          type: string
          format: date
          example: '1991-07-30'
        city:
          type: string
          example: 'Delhi'
        releasedBookedMovies:
          type: array
          items:
            type: string
          example: []
        preBookedMovies:
          type: array
          items:
            type: string
          example: []
        country:
          type: string
          example: 'India'
        zipCode:
          type: string
          example: '123456'
        state:
          type: string
          example: 'Delhi'
        status:
          type: string
          example: 'Active||Inactive'
        photo:
          type: string
          example: 'https://123.png'
        address:
          type: string
          example: '#123 5th cross Delhi'
        userType:
          type: string
          example: 'admin||user'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        firebaseId:
          type: string
          example: 'asdasd'
        passwordChangedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        passwordResetToken:
          type: string
          example: 'sadhq7ygwiueqwhe2t6783qeidasnsbhgdas'
        passwordResetExpires:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    Casts:
      type: object
      properties:
        _id:
          type: string
          format: 'hash'
          example: 'asdasd'
        firstName:
          type: string
          example: 'Jhonny'
        lastName:
          type: string
          example: 'Deep'
        photo:
          type: string
          example: 'https://photo-url.com'
        description:
          type: string
          example: 'Singh is bling'
        createdAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2023-07-30T16:10:10Z'
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: 'Something went wrong'
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes